Added layers, began groups

+++ Engine-Core/src/math/math_util.h
@@ -9,6 +9,6 @@ namespace electro_engine { namespace math {
 
 	inline float toRadians(float degrees)
 	{
-		return degrees * (M_PI / 180.0f);
+		return degrees * ((float)M_PI / 180.0f);
 	}
 } }
